+++ concurrency/CHANGELOG.rst
@@ -6,6 +6,19 @@ standard Haskell versioning scheme.
 
 .. _PVP: https://pvp.haskell.org/
 
+1.9.0.0 (2020-02-26)
+====================
+
+* Git: :tag:`concurrency-1.9.0.0`
+* Hackage: :hackage:`concurrency-1.9.0.0`
+
+Changed
+~~~~~~~
+
+* (:issue:`286`) Pulled in changes from :hackage:`stm-2.5.0.0` package:
+
+  * Changed ``newTBQueue`` to accept ``Natural`` as a size.
+  * Changed ``lengthTBQueue`` to return a ``Natural``.
 
 1.8.1.0 (2019-11-16)
 ====================
@@ -105,6 +118,8 @@ Changed
 
     * Make definition of ``readTQueue`` consistent with
       ``readTBQueue``
+
+    * Performance improvements to ``peekTQueue`` and ``peekTBQueue``.
 
 Miscellaneous
 ~~~~~~~~~~~~~